final commit

+++ bots/balthazar_farmer.py
@@ -130,8 +130,8 @@ class BotPlayer(Player):
                     tb += self.bomber_tiles[tower.x, tower.y]
                     nbombs += 1
 
-            self.desired_health = max(MINIMUM_HEALTH,
-                                      BOMBER_DAMAGE*np.ceil((BOMBER_DPS * tb)/BOMBER_DAMAGE) + GAPFILL)
+            self.desired_health = int(max(MINIMUM_HEALTH,
+                                      BOMBER_DAMAGE*np.ceil((BOMBER_DPS * tb)/BOMBER_DAMAGE) + GAPFILL))
             self.sending = self.cluster_size
 
         if self.sending > 0:
